requirements.txt
googlesearch-python
discord
aiohttp
asyncpg
psutil
timeago
emojis
dotenv
git+https://github.com/scarletcafe/jishaku@master
duckduckgo-images-api